% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{at_sea_distance}
\alias{at_sea_distance}
\title{Compute at-sea distance}
\usage{
at_sea_distance(from, to, tesselation)
}
\arguments{
\item{from, }{to
sf objects}

\item{tesselation}{An sf object representing a tesselation of the region}
}
\value{
A matrix giving the distances between from and to geometries.
}
\description{
Compute at-sea distance
}
