% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, pi, concentration, log = TRUE)
}
\arguments{
\item{x}{A probability vector}

\item{pi}{The expected probability vector}

\item{concentration}{The expected closeness of x to y}

\item{log}{Should the log density be returned?}
}
\value{
The (log) density of x
}
\description{
Dirichlet distribution
}
