% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgammamv}
\alias{dgammamv}
\title{Gamma distribution parameterized with mean and variance}
\usage{
dgammamv(x, mean, var, log = TRUE)
}
\arguments{
\item{x}{A positive numeric variable}

\item{mean}{Mean of the Gamma distribution}

\item{var}{Variance of the Gamma distribution}

\item{log}{Should the log density be returned?}
}
\description{
Gamma distribution parameterized with mean and variance
}
