% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_helpers.R
\name{lgamma}
\alias{lgamma}
\title{An RTMB-compatible version of base::lgamma}
\usage{
lgamma(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
On package-load, makes an AD-taped version of lgamma if RTMB is installed.
}
