% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_cell_key}
\alias{make_cell_key}
\title{Create a key to translate indices between rasters and tesselations}
\usage{
make_cell_key(target_raster, prediction_raster, tesselation)
}
\arguments{
\item{target_raster}{A stars raster}

\item{prediction_raster}{(Optional) A stars raster}

\item{tesselation}{(Optional) An sf data.frame containing polygons}
}
\value{
A stars raster with the same geometry as the target raster with one
or more of the following layers:
\itemize{
\item target_cell: The cell index for the target raster
\item prediction_cell: The index for the cell of the prediction raster that
coincides with the target cell
\item tesselation_cell: The index for the polygon of the tesselation that
coincides with the target cell
}
}
\description{
Create a key to translate indices between rasters and tesselations
}
