% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_directed_adjacency_graph}
\alias{make_directed_adjacency_graph}
\title{Create an adjacency matrix from a tesselation}
\usage{
make_directed_adjacency_graph(tesselation)
}
\arguments{
\item{tesselation}{An sf data.frame containing polygons}
}
\value{
A direct acyclic graph from the igraph package
}
\description{
The tesselation is converted into a directed acyclic graph
}
