% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
rdirichlet(n, pi, concentration)
}
\arguments{
\item{n}{The number of Dirichlet samples}

\item{pi}{The expected probability vector}

\item{concentration}{The expected closeness of x to y}
}
\value{
A n * length(pi) matrix
}
\description{
Dirichlet distribution
}
