% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rgammamv}
\alias{rgammamv}
\title{Gamma distribution parameterized with mean and variance}
\usage{
rgammamv(n, mean, var)
}
\arguments{
\item{n}{The number of Gamma samples}

\item{mean}{Mean of the Gamma distribution}

\item{var}{Variance of the Gamma distribution}
}
\description{
Gamma distribution parameterized with mean and variance
}
