% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_repolygonize}
\alias{st_repolygonize}
\title{Transform one polygon geometry to another using area-weighted averages}
\usage{
st_repolygonize(source, target)
}
\arguments{
\item{x}{An 'sf' or 'stars' object with POLYGON geometries}

\item{y}{An 'sf' object with the target POLYGON geometries}
}
\value{
An 'sf' or 'stars' object with the same attributes and dimensions as
x and the same geometry as y.
}
\description{
Transform one polygon geometry to another using area-weighted averages
}
