% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{tesselate}
\alias{tesselate}
\title{Make a tessellation of a polygon.}
\usage{
tesselate(shape, n, type = "hexagonal")
}
\arguments{
\item{shape}{A polygon of class sf. If multiple polygons are given, they are combined
into a single polygon before creating the tesselation.}

\item{n}{The number of polygons in the tesselation}

\item{type}{The type of tesselation. See sf::st_sample.}
}
\value{
An sf data.frame containing the tesselation polygons and an index
column.
}
\description{
Make a tessellation of a polygon.
}
