% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{crl}
\alias{crl}
\alias{icrl}
\title{Convert parameters <-> probability vector using continuation ratio logits}
\usage{
crl(qpi)

icrl(p)
}
\arguments{
\item{qpi}{An unbounded parameter vector to be transformed into a probability
vector.}

\item{p}{A probability vector to be transformed to an unbounded probability
vector.}
}
\value{
A numeric vector
}
\description{
When using crl, the output vector will have one additional element added to
the end compared to the input vector to ensure probabilities add to 1.
When using icrl, the last element of the input is discarded.
Can be vectorized, in which case elements with the same (n-1) indices
will be treated as parameters for the same probability vector.
For example if x is a 3 dimensional array, then the operation is applied to
each vector x\link{i, j, }.
}
