% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{dagify}
\alias{dagify}
\title{Create a directed adjacency matrix from a tesselation}
\usage{
dagify(tesselation)
}
\arguments{
\item{tesselation}{An sf data.frame containing polygons}
}
\value{
A direct acyclic graph from the igraph package
}
\description{
The tesselation is converted into a directed acyclic graph
}
