% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elhelpers-package.R
\docType{package}
\name{elhelpers-package}
\alias{elhelpers}
\alias{elhelpers-package}
\title{elhelpers: Helper functions used in many analyses}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ethan Lawler \email{lawlerem@gmail.com}

}
\keyword{internal}
