% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{softplus}
\alias{softplus}
\alias{isoftplus}
\title{Convert positive values to/from unbounded values.}
\usage{
softplus(x)

isoftplus(x)
}
\arguments{
\item{x}{An unbounded numeric vector}

\item{y}{A positive numeric vector}
}
\value{
A numeric vector
}
\description{
Convert positive values to/from unbounded values.
}
